move classes to fx.components
